Add files via upload
The emergence of infrastructural, operational and environmental issues such as climate change, noise pollutions, malfunctioning The has greatly augmented the need for robust, cheap, operationally adaptable, and smart monitoring systems. In this context smart sensor networks are an emerging field of research which combines many challenges of modern computer science, wireless communication and mobile computing. In this paper a solution for monitoring the noise pollution levels in the infrastructural environment using wireless embedded computing system is proposed. The solution includes the technologies that have been emerging in the field of mobile computing as well as Internet of Things (IoT) because of their vast applicability. Here, the sensing systems are connected to the embedded computing system to monitor the fluctuation of noise pollution parameters, from their normal behavior. This model is scalable and distributive for any infrastructural environment that needs continuous monitoring and behavioral analysis. Performance of the proposed model is evaluated using prototype implementation, consisting of Intel Galileo and sensor boards along with embedded programming. The implementation is tested for two to three parameters and their behavioral patterns with respect to user given specifications that provides a controlled pollution monitoring to make the environment smart.